Broken load-theme-buffer-local #5
use 'noflet' instead of 'flet' to override
the functions 'custom-theme-buffer-local-recalc-face' and
'custom-theme-buffer-local-recalc-variable'.  In later versions
of emacs, 'flet' is reverting to lexical binding instead of
dynamic binding, thus the methods are no longer being hooked
downsteam.

This introduces a dependency on package 'noflet'.  I did not
update the autoload to automatically require this package.

Without this fix, on certain versions of emacs you will get
the following message when trying to run 'load-theme-buffer-local':

Symbol's function definition is void: symbol